requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers
requests
python-multipart
jinja2